Modify: Rename a library from 'EasyChartBoard' to 'VuetifulBoard'
