+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "202108",
+  "name": "vuetiful-board",
   "version": "0.1.0",
   "private": true,
   "scripts": {
